Moving the refined implicits into its own module
